requirements.txt
numpy
pandas
matplotlib
torch
torchvision
Pillow
opencv-python
timm